Added new tests to support PDF downloads of lists and PDF downloads of summary pages for 5.7 only. Created random_methods.py file and moved all common random objet get methods into it.
